+++ R/platjam-import-metabolomics-niehs.R
@@ -6,7 +6,20 @@
 #' This import function is specific to NIEHS file formats produced
 #' from their defined analysis workflow. The files typically include
 #' `"df_pos"` for positive ionization, and `"df_neg"` for negative
-#' ionization:
+#' ionization.
+#'
+#' Optionally, when the full data processing file is present,
+#' it will be imported alongside the cleaned data described above.
+#' The full data processing imports detailed compound measurement
+#' data, and is expected in one of two formats in the `data_path` folder:
+#' 1. Files `"compounds_pos.txt"` and/or `"compounds_neg.txt"`, or
+#' 2. `"1_DataProcessed.zip"` which is expected to contain
+#' files `"compounds_pos.txt"` and/or `"compounds_neg.txt"` in the archive.
+#'
+#' The "compounds" data includes important annotations for each measurement,
+#' specifically the type of numeric measurement that is supplied
+#' by the upstream software. These annotations include whether numeric
+#' values were imputed, or measured directly in each sample.
 #'
 #' ## Sample Metadata
 #'
@@ -150,6 +163,7 @@ import_metabolomics_niehs <- function
       }
    }
 
+   #############################################
    # find metadata file
    metadata_file <- list.files(path=data_path,
       full.names=TRUE,
@@ -186,6 +200,7 @@ import_metabolomics_niehs <- function
    # now rownames should be unique
    rownames(metadata_df) <- metadata_df$CORE_Filename;
 
+   #############################################
    # find data files
    data_files <- list.files(path=data_path,
       full.names=TRUE,
@@ -404,6 +419,7 @@ import_metabolomics_niehs <- function
       return(se)
    })
 
+   #############################################
    # revise colData for each entry
    # to remove columns with no values,
    # and simplify columns with only one unique value.
@@ -492,5 +508,356 @@ import_metabolomics_niehs <- function
       return(se);
    })
 
+   # 0.0.71.900 - optionally load compound_pos.txt and compound_neg.txt
+   dp1zip <- list.files(path=data_path,
+      pattern="1_DataProcessed.zip",
+      full.names=TRUE)
+   compound_files <- list.files(path=data_path,
+      pattern="^compound_(pos|neg).txt$",
+      full.names=TRUE)
+   dp1_compound_pos <- NULL;
+   dp1_compound_neg <- NULL;
+   #
+   if (length(compound_files) < 2 && length(dp1zip) == 1) {
+      dp1files <- unzip(dp1zip, list=TRUE)
+      dp1_compound_file_pos <- jamba::vigrep("compounds_pos.txt",
+         dp1files$Name)
+      if (length(dp1_compound_file_pos) > 0) {
+         if (!file.exists(dp1_compound_file_pos)) {
+            dp1_compound_file_pos <- head(unzip(dp1zip,
+               files=dp1_compound_file_pos), 1)
+         }
+         if (verbose) {
+            jamba::printDebug("import_metabolomics_niehs(): ",
+               "Loading '", dp1_compound_file_pos,
+               "' extracted from ZIP archive.")
+         }
+         dp1_compound_pos <- data.table::fread(dp1_compound_file_pos,
+            data.table=FALSE)
+      }
+      dp1_compound_file_neg <- jamba::vigrep("compounds_neg.txt",
+         dp1files$Name)
+      if (length(dp1_compound_file_neg) > 0) {
+         if (!file.exists(dp1_compound_file_neg)) {
+            dp1_compound_file_neg <- head(unzip(dp1zip,
+               files=dp1_compound_file_neg), 1)
+         }
+         if (verbose) {
+            jamba::printDebug("import_metabolomics_niehs(): ",
+               "Loading '", dp1_compound_file_neg,
+               "' extracted from ZIP archive.")
+         }
+         dp1_compound_neg <- data.table::fread(dp1_compound_file_neg,
+            data.table=FALSE)
+      }
+   } else if (length(compound_files) > 0) {
+      # load pos file
+      cpd_file_pos <- jamba::vigrep("_pos", compound_files)
+      if (length(cpd_file_pos) == 1) {
+         if (verbose) {
+            jamba::printDebug("import_metabolomics_niehs(): ",
+               "Loading '", cpd_file_pos,
+               "'.")
+         }
+         dp1_compound_pos <- data.table::fread(cpd_file_pos,
+            data.table=FALSE)
+      }
+
+      # load neg file
+      cpd_file_neg <- jamba::vigrep("_neg", compound_files)
+      if (length(cpd_file_neg) == 1) {
+         if (verbose) {
+            jamba::printDebug("import_metabolomics_niehs(): ",
+               "Loading '", cpd_file_neg,
+               "'.")
+         }
+         dp1_compound_neg <- data.table::fread(cpd_file_neg,
+            data.table=FALSE)
+      }
+   }
+   if (length(dp1_compound_pos) > 0) {
+      if (verbose) {
+         jamba::printDebug("import_metabolomics_niehs(): ",
+            "Merging compounds_pos into se_list.")
+      }
+      se_list <- process_metab_compounds_file(dp1_compound_pos,
+         type="pos",
+         se_list=se_list)
+   }
+   if (length(dp1_compound_neg) > 0) {
+      if (verbose) {
+         jamba::printDebug("import_metabolomics_niehs(): ",
+            "Merging compounds_neg into se_list.")
+      }
+      se_list <- process_metab_compounds_file(dp1_compound_neg,
+         type="neg",
+         se_list=se_list)
+   }
+
+   # return SE list
    return(se_list)
+}
+
+#' Process NIEHS compounds output files (internal)
+#'
+#' Process NIEHS compounds output files (internal)
+#'
+#' This function produces `SummarizedExperiment` object with `assays`
+#' defined by the `assay_names_extra` patterns below. Each assay name
+#' is the lowercase value, with spaces replaced with underscore `"_"`.
+#' For example `"Gap Fill Status"` will become assay name `"gap_fill_status"`.
+#'
+#' @family jam import functions
+#'
+#' @param x `data.frame` after reading `compound_pos.txt` or `compound_neg.txt`
+#' @param nsmall_mz `numeric` argument passed to `format()` to define
+#'    the number of decimal positions to retain in the output label
+#'    from column `"m/z"`.
+#' @param nsmall_rt `numeric` argument passed to `format()` to define
+#'    the number of decimal positions to retain in the output label
+#'    from column `"RT [min]"`.
+#' @param assay_names_extra `character` vector with prefix values expected
+#'    in column names, for example where `"Area"` is expected to form a prefix
+#'    to column names: `"Area: sample_a", "Area: sample_b"`;
+#'    and `"Gap Fill Status"` would form column names
+#'    `"Gap Fill Status: sample_a", "Gap Fill Status: sample_b"`.
+#' @param mz_colname,rt_colname `character` string to indicate the appropriate
+#'    column name containing `"m/z"` (mass/charge) and `"RT [min]"`
+#'    (retention time in minutes) as used to format a unique identifier
+#'    for each row.
+#' @param type `character` string indicating whether data represents
+#'    positive `"pos"` or negative `"neg"` ionization. Mainly used
+#'    when `se_list` is also provided.
+#' @param se_list optional `list` of `SummarizedExperiment` objects,
+#'    named with `"pos"` or `"neg"` in the name. When provided, the
+#'    `se` output of this function is merged into the `se_list` with
+#'    matching `type`, adding `assay_names` into the corresponding
+#'    `se_list` object.
+#' @param ... additional arguments are ignored.
+#'
+#' @returns `SummarizedExperiment` object with `assays` defined by matching
+#'    column names in `x`. When no column names meet the `assay_names_extra`
+#'    criteria, this function returns `NULL`.
+#'
+#' @export
+process_metab_compounds_file <- function
+(x,
+ nsmall_mz=4,
+ nsmall_rt=2,
+ assay_names_extra=c(
+    "Area",
+    "Gap Fill Status",
+    "Peak Rating"),
+ mz_colname="m/z",
+ rt_colname="RT [min]",
+ type=c("pos",
+    "neg"),
+ se_list=NULL,
+ ...)
+{
+   type <- match.arg(type);
+
+   # extract annotation and assay data
+   names(assay_names_extra) <- gsub("[ ]+",
+      "_",
+      tolower(assay_names_extra))
+   non_assay_colnames <- jamba::unvigrep(
+      paste0("^",
+         jamba::cPaste(assay_names_extra, "|"),
+         ":"),
+      colnames(x))
+   # head(x[, non_assay_colnames], 5)
+
+   x_names <- paste0(
+      seq_len(nrow(x)),
+      "|",
+      format(
+         round(x[, mz_colname]*(10^nsmall_mz))/(10^nsmall_mz),
+         nsmall=nsmall_mz,
+         trim=TRUE),
+      "|",
+      format(
+         round(x[, rt_colname]*(10^nsmall_rt))/(10^nsmall_rt),
+         nsmall=nsmall_rt,
+         trim=TRUE))
+   # print(jamba::vigrep("^(27|12)[|]", x_names));
+   rownames(x) <- x_names;
+
+   # prepare rowData
+   x_rowData <- x[, non_assay_colnames, drop=FALSE];
+
+   # assays
+   assays_list <- lapply(assay_names_extra, function(iassay_name){
+      iassay_colnames <- jamba::vigrep(
+         paste0("^", iassay_name, ":"), colnames(x))
+      if (length(iassay_colnames) == 0) {
+         return(NULL)
+      }
+      ix <- x[, iassay_colnames, drop=FALSE]
+      # remove prefix
+      colnames(ix) <- gsub(paste0("^", iassay_name, ":[ ]*"),
+         "",
+         colnames(ix))
+      # remove suffix
+      colnames(ix) <- gsub("[ ]*[(][A-Z0-9]+[)][ ]*$",
+         "",
+         colnames(ix))
+      # remove suffix
+      colnames(ix) <- gsub("[.](raw)$",
+         "",
+         colnames(ix))
+      ix <- as.matrix(ix)
+      if (max(ix, na.rm=TRUE) > 1000) {
+         ix <- log2(1 + ix);
+      }
+      ix
+   })
+   # remove empty entries
+   assays_list <- jamba::rmNULL(assays_list);
+   if (length(assays_list) == 0) {
+      jamba::printDebug("process_metab_compounds_file(): ",
+         "No matching assay names were recognized: ",
+         paste0("'", assay_names_extra, "'"),
+         ". Returning NULL.")
+      return(se_list)
+   }
+   se <- SummarizedExperiment::SummarizedExperiment(
+      assays=assays_list,
+      rowData=x_rowData)
+
+   # now merge with incoming se
+   # match rows to se_list
+   if (length(se_list) > 0) {
+      se_names <- jamba::vigrep(type, names(se_list));
+      for (se_name in se_names) {
+         # match by rownum only - no other matching is reliable
+         match_pos_rows <- match(
+            gsub("[|].+$", "",
+               rownames(se_list[[se_name]])),
+            gsub("[|].+$", "",
+               rownames(se)))
+         rownames(se)[match_pos_rows] <- rownames(se_list[[se_name]])
+         match_pos_columns <- match(colnames(se_list[[se_name]]),
+            colnames(se))
+         for (iassay_name in SummarizedExperiment::assayNames(se)) {
+            SummarizedExperiment::assays(
+               se_list[[se_name]])[[iassay_name]] <- (
+                  SummarizedExperiment::assays(
+                     se[match_pos_rows, match_pos_columns])[[iassay_name]])
+         }
+      }
+      return(se_list)
+   }
+
+   return(se)
+}
+
+
+#' Convert imputed metabolomics assay data to NA
+#'
+#' Convert imputed metabolomics assay data to NA
+#'
+#' This function takes `SummarizedExperiment` data as input, which
+#' must contain at least two `SummarizedExperiment::assayNames()`
+#' entries:
+#' 1. `numeric` assay measurements
+#' 2. `numeric` impute flag data
+#'
+#' The coordinates of cells in the impute flag data
+#' which match `impute_flags` are used to convert data in assay measurements
+#' to `NA`.
+#'
+#' @family jam utility functions
+#'
+#' @returns `SummarizedExperiment` or when input `se` is a `list`
+#'    a `list` of `SummarizedExperiment` objects is returned.
+#'
+#' @param se `SummarizedExperiment` object, or `list` of
+#'    `SummarizedExperiment` objects.
+#'    When a `list` is provided, each entry is iterated and processed,
+#'    and a `list` is returned.
+#' @param measurement_assay_names `character` vector with one or more
+#'    `SummarizedExperiment::assayNames(se)` for which the imputed
+#'    data will be converted to `NA` values.
+#' @param impute_assay_name `character` string with one
+#'    `SummarizedExperiment::assayNames(se)` entry that contains the
+#'    impute flags. Only the first matching assay name is used.
+#' @param suffix `character` suffix appended to each value in
+#'    `measurement_assay_names` to create a new assay name for the
+#'    resulting data matrix.
+#'
+#'    When filtering by different impute criteria, it may be helpful
+#'    to include the codes in the suffix, for example filtering by only
+#'    `impute_flags=8` may suggest `suffix="_noimpute8"`.
+#' @param impute_flags `numeric` vector of flags indicating values in
+#'    `impute_assay_name` assay data which represent imputed data.
+#'
+#'    The defaults:
+#'    * 8: filled by trace area (imputed)
+#'    * 32: filled by spectrum noise (imputed)
+#'
+#'    Other values:
+#'    * 0: original ion used
+#'    * 64: filled by matching ion
+#'    * 128: filled by re-detected peak
+#' @param ... additional arguments are ignored.
+#'
+#' @export
+convert_imputed_assays_to_na <- function
+(se,
+ measurement_assay_names="area",
+ impute_assay_name="gap_fill_status",
+ suffix="_noimpute",
+ impute_flags=c(8, 32),
+ ...)
+{
+   #
+   if (is.list(se)) {
+      for (i in seq_along(se)) {
+         se[[i]] <- convert_imputed_assays_to_na(se[[i]],
+            measurement_assay_names=measurement_assay_names,
+            impute_assay_name=impute_assay_name,
+            suffix=suffix,
+            impute_flags=impute_flags,
+            ...)
+      }
+      return(se)
+   }
+
+   if (length(impute_flags) == 0) {
+      stop("No impute_flags were defined.")
+   }
+
+   use_impute_assay_name <- head(intersect(impute_assay_name,
+      SummarizedExperiment::assayNames(se)), 1)
+   if (length(use_impute_assay_name) == 0) {
+      stop(paste0("impute_assay_name '",
+         impute_assay_name,
+         "' was not found in assayNames(se)."))
+   }
+
+   use_assay_names <- intersect(measurement_assay_names,
+      SummarizedExperiment::assayNames(se))
+   if (length(use_assay_names) == 0) {
+      stop(paste0("No measurement_assay_names (",
+         jamba::cPaste(paste0("'", measurement_assay_names, "'")),
+         ") were present in ",
+         "assayNames(se)."))
+   }
+
+   # define imputed data positions
+   is_imputed <- SummarizedExperiment::assays(se)[[use_impute_assay_name]]
+   is_imputed[] <- (SummarizedExperiment::assays(se)[[use_impute_assay_name]]
+      %in% c(impute_flags))
+
+   for (use_assay_name in use_assay_names) {
+      new_assay_name <- paste0(use_assay_name, suffix)
+      SummarizedExperiment::assays(se)[[new_assay_name]] <- (
+         SummarizedExperiment::assays(se)[[use_assay_name]]);
+      if (any(is_imputed %in% 1)) {
+         SummarizedExperiment::assays(se)[[
+            new_assay_name]][is_imputed %in% 1] <- NA;
+      }
+   }
+   return(se);
 }